Add support for special float literals (#1965)
* Add support for infinite default values

Both the proto2 spec [1] and proto3 spec [2] define floating-point literal values "inf" and "-inf"
that may be used as default values for both float and double fields. This change adds support for
such literal values in both the Java and Kotlin generators

[1] https://developers.google.com/protocol-buffers/docs/reference/proto2-spec#floating-point-literals
[2] https://developers.google.com/protocol-buffers/docs/reference/proto3-spec#floating-point-literals

* Add support for NaN default values

Both the proto2 spec [1] and proto3 spec [2] define the floating-point literal values "nan" that
may be used as default values for both float and double fields. This change adds support for such
literal values in both the Java and Kotlin generators.

[1] https://developers.google.com/protocol-buffers/docs/reference/proto2-spec#floating-point-literals
[2] https://developers.google.com/protocol-buffers/docs/reference/proto3-spec#floating-point-literals