Handle writing/reading exceptions for duplex calls (#2038)
* Handle writing/reading exceptions for duplex calls

For reading, an exception would be thrown when we try to read a message
when the stream had already been closed. We now properly handle all
exceptions and pass them along into our sending channel.

For sending, even though flushing the message would have failed, the
thrown exception would be eaten up and we would not realize the writing
has failed from upstream. We now catch it and pass it along into our
receive channel.

* More defense against close failures

* Test duplexSuspend_requestChannelThrowsWhenWhenCanceledByServer

Co-authored-by: Jesse Wilson <jwilson@squareup.com>