Add the proto source file to the message adapter
This PR adds a property to the Wire message adapter to point at the message's proto file path.

This change will be used to integrate Wire with Confluent's Schema Registry. Confluent's protobuf implementation uses the encoded message descriptors in the generated Java code. We need similar functionality in Wire. In order to do that  we need to access the proto file for a message. Wire already supports including the proto file in the classpath resources via ```protoLibrary = true``` configuration option. This change simplifies discovering the proto file for a message by adding it to the adapter.

Checks:
- [X] Code compiles correctly
- [X] Updated ```JavaGeneratorTest.generateAbstractAdapter``` which failed with this new change.