Copy a specific compiler artifact in Swift
Previously the copy would fail if there happened to be multiple artifacts in the build directory because `cp` with a glob-star fails if it matches multiple sources with a single destination. Instead we now copy the most recently modified artifact. We don't use the `version` variable because it's possible that the Podspec's version is different from the version of the Gradle artifact.